Split the speech.py file into separate `SpeechCog` and `TTSController` single class files
- Updated speech imports to reference the new `SpeechCog`
- Hoisted the `TTSController`'s cwd setter logic out to `hawking.py`
- Converted all format strings to f-strings in `TTSController`
